requirements.txt
# --extra-index-url https://download.pytorch.org/whl/cu113
# clip @ git+https://github.com/openai/CLIP.git
chumpy
decord==0.6.0
einops==0.6.1
ffmpy==0.3.1
ftfy==6.1.1
gdown==4.7.1
grpcio==1.54.2
h11==0.14.0
importlib-metadata==5.0.0
importlib-resources==5.12.0
joblib
matplotlib==3.3.1 
numpy==1.21.5
Pillow==9.2.0
PyYAML==6.0
scikit-learn
scipy 
smplx==0.1.28
sniffio==1.3.0
timm==0.9.7
# torch==1.12.0
torch-tb-profiler
# torchaudio
# torchvision
tornado
tqdm
trimesh
vector-quantize-pytorch==1.6.30
